// packages/oos-pipeline-ts/src/generated/ast.ts
export interface AndExpr extends langium.AstNode {
    readonly $container: AndExpr | NotExpr | OrExpr | StepSemantic | StepWhere;
    readonly $type: 'AndExpr';
    exprs: Array<WhereExpr>;
}

export const AndExpr = {
    $type: 'AndExpr',
    exprs: 'exprs'
} as const;

export function isAndExpr(item: unknown): item is AndExpr {
    return reflection.isInstance(item, AndExpr.$type);
}

export interface EmbeddingDecl extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'EmbeddingDecl';
    model: string;
}

export const EmbeddingDecl = {
    $type: 'EmbeddingDecl',
    model: 'model'
} as const;

export function isEmbeddingDecl(item: unknown): item is EmbeddingDecl {
    return reflection.isInstance(item, EmbeddingDecl.$type);
}

export interface FieldFilter extends langium.AstNode {
    readonly $container: AndExpr | NotExpr | OrExpr | StepSemantic | StepWhere;
    readonly $type: 'FieldFilter';
    field: string;
    op: FilterOp;
    value: FilterValue;
}

export const FieldFilter = {
    $type: 'FieldFilter',
    field: 'field',
    op: 'op',
    value: 'value'
} as const;

export function isFieldFilter(item: unknown): item is FieldFilter {
    return reflection.isInstance(item, FieldFilter.$type);
}

export type FilterOp = '_eq' | '_gt' | '_gte' | '_ilike' | '_in' | '_like' | '_lt' | '_lte' | '_neq';

export function isFilterOp(item: unknown): item is FilterOp {
    return item === '_eq' || item === '_neq' || item === '_gt' || item === '_gte' || item === '_lt' || item === '_lte' || item === '_like' || item === '_ilike' || item === '_in';
}

export type FilterValue = NumberValue | ParamRef | StringValue;

export const FilterValue = {
    $type: 'FilterValue'
} as const;

export function isFilterValue(item: unknown): item is FilterValue {
    return reflection.isInstance(item, FilterValue.$type);
}

export interface LlmDecl extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'LlmDecl';
    model: string;
}

export const LlmDecl = {
    $type: 'LlmDecl',
    model: 'model'
} as const;

export function isLlmDecl(item: unknown): item is LlmDecl {
    return reflection.isInstance(item, LlmDecl.$type);
}

export interface ModeDecl extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'ModeDecl';
    value: PipelineMode;
}

export const ModeDecl = {
    $type: 'ModeDecl',
    value: 'value'
} as const;

export function isModeDecl(item: unknown): item is ModeDecl {
    return reflection.isInstance(item, ModeDecl.$type);
}

export interface NotExpr extends langium.AstNode {
    readonly $container: AndExpr | NotExpr | OrExpr | StepSemantic | StepWhere;
    readonly $type: 'NotExpr';
    expr: WhereExpr;
}

export const NotExpr = {
    $type: 'NotExpr',
    expr: 'expr'
} as const;

export function isNotExpr(item: unknown): item is NotExpr {
    return reflection.isInstance(item, NotExpr.$type);
}

export interface NumberValue extends langium.AstNode {
    readonly $container: FieldFilter;
    readonly $type: 'NumberValue';
    value: number;
}

export const NumberValue = {
    $type: 'NumberValue',
    value: 'value'
} as const;

export function isNumberValue(item: unknown): item is NumberValue {
    return reflection.isInstance(item, NumberValue.$type);
}

export interface OrExpr extends langium.AstNode {
    readonly $container: AndExpr | NotExpr | OrExpr | StepSemantic | StepWhere;
    readonly $type: 'OrExpr';
    exprs: Array<WhereExpr>;
}

export const OrExpr = {
    $type: 'OrExpr',
    exprs: 'exprs'
} as const;

export function isOrExpr(item: unknown): item is OrExpr {
    return reflection.isInstance(item, OrExpr.$type);
}

export interface OutDecl extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'OutDecl';
    subject?: string;
    target: OutTarget;
}

export const OutDecl = {
    $type: 'OutDecl',
    subject: 'subject',
    target: 'target'
} as const;

export function isOutDecl(item: unknown): item is OutDecl {
    return reflection.isInstance(item, OutDecl.$type);
}

export type OutTarget = 'editor' | 'file' | 'nats';

export function isOutTarget(item: unknown): item is OutTarget {
    return item === 'editor' || item === 'nats' || item === 'file';
}

export interface ParamRef extends langium.AstNode {
    readonly $container: FieldFilter;
    readonly $type: 'ParamRef';
    param: string;
}

export const ParamRef = {
    $type: 'ParamRef',
    param: 'param'
} as const;

export function isParamRef(item: unknown): item is ParamRef {
    return reflection.isInstance(item, ParamRef.$type);
}

export interface PipelineDef extends langium.AstNode {
    readonly $type: 'PipelineDef';
    embedding?: EmbeddingDecl;
    llm: LlmDecl;
    mode: ModeDecl;
    name: string;
    out: OutDecl;
    sources: Array<Source>;
    steps: Array<Step>;
}

export const PipelineDef = {
    $type: 'PipelineDef',
    embedding: 'embedding',
    llm: 'llm',
    mode: 'mode',
    name: 'name',
    out: 'out',
    sources: 'sources',
    steps: 'steps'
} as const;

export function isPipelineDef(item: unknown): item is PipelineDef {
    return reflection.isInstance(item, PipelineDef.$type);
}

export type PipelineMode = 'context' | 'mass';

export function isPipelineMode(item: unknown): item is PipelineMode {
    return item === 'context' || item === 'mass';
}

export interface SemanticQuery extends langium.AstNode {
    readonly $container: StepSemantic;
    readonly $type: 'SemanticQuery';
    value: string;
}

export const SemanticQuery = {
    $type: 'SemanticQuery',
    value: 'value'
} as const;

export function isSemanticQuery(item: unknown): item is SemanticQuery {
    return reflection.isInstance(item, SemanticQuery.$type);
}

export type Source = SourceDb | SourceDocs;

export const Source = {
    $type: 'Source'
} as const;

export function isSource(item: unknown): item is Source {
    return reflection.isInstance(item, Source.$type);
}

export interface SourceDb extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'SourceDb';
    dsn?: string;
    name: string;
    table: string;
}

export const SourceDb = {
    $type: 'SourceDb',
    dsn: 'dsn',
    name: 'name',
    table: 'table'
} as const;

export function isSourceDb(item: unknown): item is SourceDb {
    return reflection.isInstance(item, SourceDb.$type);
}

export interface SourceDocs extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'SourceDocs';
    name: string;
    uri: string;
}

export const SourceDocs = {
    $type: 'SourceDocs',
    name: 'name',
    uri: 'uri'
} as const;

export function isSourceDocs(item: unknown): item is SourceDocs {
    return reflection.isInstance(item, SourceDocs.$type);
}

export type Step = StepLlm | StepSemantic | StepWhere;

export const Step = {
    $type: 'Step'
} as const;

export function isStep(item: unknown): item is Step {
    return reflection.isInstance(item, Step.$type);
}

export interface StepLlm extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'StepLlm';
    input?: langium.Reference<Step>;
    mode: StepMode;
    name: string;
    prompt: string;
    source: string;
    system?: string;
}

export const StepLlm = {
    $type: 'StepLlm',
    input: 'input',
    mode: 'mode',
    name: 'name',
    prompt: 'prompt',
    source: 'source',
    system: 'system'
} as const;

export function isStepLlm(item: unknown): item is StepLlm {
    return reflection.isInstance(item, StepLlm.$type);
}

export type StepMode = 'aggregate' | 'per-row';

export function isStepMode(item: unknown): item is StepMode {
    return item === 'per-row' || item === 'aggregate';
}

export interface StepSemantic extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'StepSemantic';
    limit?: number;
    name: string;
    queries: Array<SemanticQuery>;
    source: string;
    where?: WhereExpr;
}

export const StepSemantic = {
    $type: 'StepSemantic',
    limit: 'limit',
    name: 'name',
    queries: 'queries',
    source: 'source',
    where: 'where'
} as const;

export function isStepSemantic(item: unknown): item is StepSemantic {
    return reflection.isInstance(item, StepSemantic.$type);
}

export interface StepWhere extends langium.AstNode {
    readonly $container: PipelineDef;
    readonly $type: 'StepWhere';
    limit?: number;
    name: string;
    source: langium.Reference<SourceDb>;
    where?: WhereExpr;
}

export const StepWhere = {
    $type: 'StepWhere',
    limit: 'limit',
    name: 'name',
    source: 'source',
    where: 'where'
} as const;

export function isStepWhere(item: unknown): item is StepWhere {
    return reflection.isInstance(item, StepWhere.$type);
}

export interface StringValue extends langium.AstNode {
    readonly $container: FieldFilter;
    readonly $type: 'StringValue';
    value: string;
}

export const StringValue = {
    $type: 'StringValue',
    value: 'value'
} as const;

export function isStringValue(item: unknown): item is StringValue {
    return reflection.isInstance(item, StringValue.$type);
}

export type WhereExpr = AndExpr | FieldFilter | NotExpr | OrExpr;

export const WhereExpr = {
    $type: 'WhereExpr'
} as const;